update reqs.txt, rename parsevariables
