+++ worker/requirements.txt
@@ -11,7 +11,7 @@ idna==2.8
 importlib-metadata==1.4.0
 importlib-resources==1.0.2
 jmespath==0.9.4
-more-itertools==8.1.0
+more-itertools==5.0.0
 nodeenv==1.3.4
 pandas==0.24.2
 pre-commit==1.21.0
