+++ worker/step_sim/step_osm/parse_variables.py
